requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.1
django-braces==1.14.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-restframework==0.0.1
djangorestframework==3.11.1
firebase-admin==4.3.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
mysql-connector-python==8.0.21
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
